// backend/src/modules/patient-exercises/interfaces/patient-exercise-status.enum.ts
// src/modules/patient-exercises/interfaces/patient-exercise-status.enum.ts
export enum PatientExerciseStatus {
  ASSIGNED = 'assigned',
  IN_PROGRESS = 'in-progress',
  COMPLETED = 'completed',
  OVERDUE = 'overdue',
  CANCELLED = 'cancelled',
}

export enum PriorityLevel {
  LOW = 'low',
  MEDIUM = 'medium',
  HIGH = 'high',
}